backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
werkzeug
pyjwt
click